corrected the folder structure
