--- services/cms/backend/app.module.ts
@@ -1,21 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ConfigModule } from '@nestjs/config';
-import { MongooseModule } from '@nestjs/mongoose';
-import { UsersModule } from './users/users.module';
-import { JobsModule } from './jobs/jobs.module';
-import { AuthModule } from './auth/auth.module';
-import { InterviewsModule } from './interviews/interviews.module';
-import { ApplicationsModule } from './applications/applications.module';
-
-@Module({
-  imports: [
-    ConfigModule.forRoot({ isGlobal: true }),
-    MongooseModule.forRoot(process.env.MONGO_URI),
-    AuthModule,
-    UsersModule,
-    JobsModule,
-    InterviewsModule,
-    ApplicationsModule,
-  ],
-})
-export class AppModule {}
