--- services/cms/backend/applications/applications.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import { HttpModule } from '@nestjs/axios';
-import { ApplicationsController } from './applications.controller';
-import { ApplicationsService } from './applications.service';
-
-@Module({
-  imports: [HttpModule],
-  controllers: [ApplicationsController],
-  providers: [ApplicationsService],
-  exports: [ApplicationsService],
-})
-export class ApplicationsModule {}
