--- services/cms/backend/jobs/jobs.module.ts
@@ -1,15 +1,0 @@
-// src/jobs/jobs.module.ts
-import { Module } from '@nestjs/common';
-import { MongooseModule } from '@nestjs/mongoose';
-import { JobsService } from './jobs.service';
-import { JobsController } from './jobs.controller';
-import { AuthModule } from 'src/auth/auth.module';
-import { HttpModule } from '@nestjs/axios';
-
-@Module({
-  imports: [HttpModule],
-  controllers: [JobsController],
-  providers: [JobsService],
-  exports: [JobsService],
-})
-export class JobsModule {}
