--- services/cms/backend/src/applications/applications.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ApplicationsController } from './applications.controller';
-import { ApplicationsService } from './applications.service';
-
-describe('ApplicationsController', () => {
-  let controller: ApplicationsController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [ApplicationsController],
-      providers: [ApplicationsService],
-    }).compile();
-
-    controller = module.get<ApplicationsController>(ApplicationsController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
