+++ services/cms/backend/src/applications/applications.module.ts
@@ -1,9 +1,12 @@
 import { Module } from '@nestjs/common';
-import { ApplicationsService } from './applications.service';
+import { HttpModule } from '@nestjs/axios';
 import { ApplicationsController } from './applications.controller';
+import { ApplicationsService } from './applications.service';
 
 @Module({
+  imports: [HttpModule],
   controllers: [ApplicationsController],
   providers: [ApplicationsService],
+  exports: [ApplicationsService],
 })
 export class ApplicationsModule {}
