--- services/cms/backend/src/applications/applications.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ApplicationsService } from './applications.service';
-
-describe('ApplicationsService', () => {
-  let service: ApplicationsService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [ApplicationsService],
-    }).compile();
-
-    service = module.get<ApplicationsService>(ApplicationsService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
