--- services/cms/backend/src/applications/dto/create-application.dto.ts
@@ -1,1 +1,0 @@
-export class CreateApplicationDto {}
