--- services/cms/backend/src/applications/dto/update-application.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateApplicationDto } from './create-application.dto';
-
-export class UpdateApplicationDto extends PartialType(CreateApplicationDto) {}
