--- services/cms/backend/src/applications/entities/application.entity.ts
@@ -1,1 +1,0 @@
-export class Application {}
