--- services/cms/backend/src/jobs/dto/create-job.dto.ts
@@ -1,14 +1,0 @@
-import { IsString, IsOptional, IsArray } from 'class-validator';
-
-export class CreateJobDto {
-  @IsString()
-  title: string;
-
-  @IsOptional()
-  @IsString()
-  description?: string;
-
-  @IsOptional()
-  @IsArray()
-  skills?: string[];
-}
