--- services/cms/backend/src/jobs/dto/update-job.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateJobDto } from './create-job.dto';
-
-export class UpdateJobDto extends PartialType(CreateJobDto) {}
