--- services/cms/backend/src/jobs/jobs.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { JobsController } from './jobs.controller';
-import { JobsService } from './jobs.service';
-
-describe('JobsController', () => {
-  let controller: JobsController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [JobsController],
-      providers: [JobsService],
-    }).compile();
-
-    controller = module.get<JobsController>(JobsController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
