+++ services/cms/backend/src/jobs/jobs.module.ts
@@ -3,14 +3,11 @@ import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 import { JobsService } from './jobs.service';
 import { JobsController } from './jobs.controller';
-import { Job, JobSchema } from './schemas/job.schema';
 import { AuthModule } from 'src/auth/auth.module';
+import { HttpModule } from '@nestjs/axios';
 
 @Module({
-  imports: [
-    MongooseModule.forFeature([{ name: Job.name, schema: JobSchema }]),
-    AuthModule
-  ],
+  imports: [HttpModule],
   controllers: [JobsController],
   providers: [JobsService],
   exports: [JobsService],
