--- services/cms/backend/src/jobs/jobs.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { JobsService } from './jobs.service';
-
-describe('JobsService', () => {
-  let service: JobsService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [JobsService],
-    }).compile();
-
-    service = module.get<JobsService>(JobsService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
