--- services/cms/backend/users/dto/create-user.dto.ts
@@ -1,22 +1,0 @@
-// src/users/dto/create-user.dto.ts
-import { IsEmail, IsString, IsOptional, IsEnum } from 'class-validator';
-import { UserRole } from '../schemas/user.schema';
-
-export class CreateUserDto {
-  @IsEmail()
-  email: string;
-
-  @IsString()
-  password: string;
-
-  @IsEnum(UserRole)
-  role: UserRole; // 'admin' or 'hr'
-
-  @IsString()
-  @IsOptional()
-  firstName?: string;
-
-  @IsString()
-  @IsOptional()
-  lastName?: string;
-}
