--- services/cms/backend/users/dto/update-user.dto.ts
@@ -1,15 +1,0 @@
-import { IsString, IsOptional } from 'class-validator';
-
-export class UpdateUserDto {
-  @IsOptional()
-  @IsString()
-  password?: string;
-
-  @IsOptional()
-  @IsString()
-  firstName?: string;
-
-  @IsOptional()
-  @IsString()
-  lastName?: string;
-}
